Code Work/research-assistant/requirements.txt
langgraph
langchain
langchain-ollama
langchain-tavily
reportlab
streamlit
rich
python-dotenv
